v1.1.0 to fix `multiple-definitions` linker error
### Releases v1.1.0

1. Fix `multiple-definitions` linker error. Drop `src_cpp` and `src_h` directories

+++ examples/RPM_Measure/RPM_Measure.ino
@@ -73,8 +73,10 @@ ESP32Timer ITimer0(0);
 volatile unsigned long rotationTime = 0;
 
 // Not using float => using RPM = 100 * real RPM
-uint32_t RPM       = 0;
-uint32_t avgRPM    = 0;
+float RPM       = 0;
+float avgRPM    = 0;
+//uint32_t RPM       = 0;
+//uint32_t avgRPM    = 0;
 
 volatile int debounceCounter;
